Move string_utils.hpp into its own directory.
