+++ components/core/src/Grep.cpp
@@ -8,7 +8,7 @@
 #include "ir/parsing.hpp"
 #include "ir/types.hpp"
 #include "LogSurgeonReader.hpp"
-#include "string_utils.hpp"
+#include "string_utils/string_utils.hpp"
 #include "StringReader.hpp"
 #include "Utils.hpp"
