+++ components/core/src/ffi/search/CompositeWildcardToken.cpp
@@ -2,7 +2,7 @@
 
 #include "../../ir/parsing.hpp"
 #include "../../ir/types.hpp"
-#include "string_utils.hpp"
+#include "string_utils/string_utils.hpp"
 
 using std::string;
 using std::string_view;
