+++ components/core/src/ffi/search/query_methods.cpp
@@ -4,7 +4,7 @@
 #include "../../ir/types.hpp"
 #include "CompositeWildcardToken.hpp"
 #include "QueryMethodFailed.hpp"
-#include "string_utils.hpp"
+#include "string_utils/string_utils.hpp"
 
 using ir::eight_byte_encoded_variable_t;
 using ir::four_byte_encoded_variable_t;
