+++ components/core/tests/test-string_utils.cpp
@@ -4,7 +4,7 @@
 #include <boost/range/combine.hpp>
 #include <Catch2/single_include/catch2/catch.hpp>
 
-#include "string_utils.hpp"
+#include "string_utils/string_utils.hpp"
 
 using std::cout;
 using std::endl;
